# docs/api/User_Api_Swagger.yaml
...
openapi: 3.0.0
info:
  title: ESP RainMaker Neo Auth & Accounts API
  version: 0.8.0
  description: |
    Sign-in and account management for the ESP RainMaker Neo cloud.
    Anything that produces or rotates an access token lives here; once
    you have a token, the rest of the platform lives in the
    **ESP RainMaker Neo Platform API**

    ESP User is the platform's own OAuth 2.1 / OIDC issuer: every token a client
    holds is minted here, regardless of how the user proved their identity.
    Admins authenticate directly against their AWS Cognito user pool.

    Login methods (all end in OUR tokens; upstream credentials never reach clients):

    - **Federated login (default)**: `GET /oauth2/authorize` starts a browser
      authorization-code + PKCE flow. With a single enabled identity provider the
      user is auto-redirected to it; the
      broker exchanges the upstream code server-side and issues OUR authorization
      code at the client's `redirect_uri`.
    - **Native password APIs** (`/v1/user/auth/*`): the main-era
      native surface. Credentials are verified against the bundled Cognito pool server-side and converted to our tokens.

    Token lifecycle after login:

    1. `POST /oauth2/token` (`grant_type=refresh_token`) - rotate the token set
       (native `/v1/user/auth/token/refresh` is reuse-tolerant instead).
    2. Send `access_token` as a `Bearer` `Authorization` header on subsequent
...
       must come from the same sign-in.
...
  contact:
...
  license:
...

...
servers:
...
paths:
...
  /oauth2/authorize:
    get:
      summary: Start browser login (authorization code + PKCE)
      description: |
        Starts the authorization-code + PKCE (`S256`) login. Validates the client, the exact `redirect_uri` and PKCE, sets the `__Host-esp_flow_id` cookie, and sends the browser to the login: straight to `/oauth2/federation/start` when exactly one provider is enabled, to the OTP page in an OTP-only deployment, otherwise to `/oauth2/login`. On success the browser returns to `redirect_uri` with a single-use `code` and `state`.

        A live `__Host-esp_session` cookie skips the login and returns the code immediately (single sign-on), unless `prompt`, `max_age` or `id_token_hint` rule it out.
      tags:
        - OAuth2 / OIDC Core
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: ["code"]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          description:
            Must exactly match a registered redirect URI for the client.
          schema:
            type: string
        - name: scope
          in: query
          required: false
          description:
            Optional (RFC 6749 §4.1.1), space-delimited; include `openid` for an
            id_token. Each requested scope must be within the client's allowed
            scopes.
          schema:
            type: string
            example: "openid email"
        - name: code_challenge
          in: query
          required: false
          description:
            PKCE code challenge (RFC 7636). Required for clients registered with
            `require_pkce` (always the case for public clients); optional
            otherwise.
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: false
          description: Must be `S256` when a `code_challenge` is provided.
          schema:
            type: string
            enum: ["S256"]
        - name: state
          in: query
          required: false
          description:
            Opaque value echoed back on the redirect (client CSRF token).
          schema:
            type: string
        - name: resource
          in: query
          required: false
          description: |
            RFC 8707 resource indicator: the API the access token is for, stamped into
            `aud` on that token and on every token the resulting refresh family later
            mints. Must appear in the client's `allowed_resources`, or the request is
            rejected with `invalid_target`.

            **Default when omitted:** `aud` is the `client_id`. **At most one** value; two
            are refused rather than minting a token good at two APIs. An ID token's `aud`
            is the `client_id` either way — it is a statement to that client.
          schema:
            type: string
            format: uri
          example: "https://api.accounts.espressif.com"
        - name: prompt
          in: query
          required: false
          description: |
            OIDC Core §3.1.2.1. **Default when
            omitted:** a live session answers the request silently.

            `none` never shows a prompt — it answers from the session or redirects back
            with `error=login_required`. `login` forces re-authentication even with a live
            session, and is carried to the upstream provider so it re-authenticates too.

            `consent` and `select_account` are accepted, but this server does not yet
            distinguish them from `login`: all three disable the session shortcut, and the
            value is forwarded verbatim to the upstream provider. Neither renders a surface
            of its own here — there is no consent screen, and account choice belongs to the
            provider — so any difference you observe between them comes from the provider,
            not from us.
          schema:
            type: string
            example: "none"
        - name: max_age
          in: query
          required: false
          description: |
            Maximum acceptable age, in seconds, of the session's authentication (its
            `auth_time`, inherited from the upstream provider). A session older than this
            is not used and the user re-authenticates.

            **Default when omitted:** no age bound — a session is used however old its
            authentication, up to the session's own idle and absolute limits. `max_age=0`
            always forces re-authentication. Carried to the upstream provider alongside
            `prompt`.
          schema:
            type: integer
            minimum: 0
        - name: id_token_hint
          in: query
          required: false
          description: |
            A previously issued ID token naming the expected subject. A session
            belonging to anyone else — or a hint that does not verify — is not
            used; the flow never switches to the hinted user.
          schema:
            type: string
      responses:
        "302":
          description: |
            Redirect to the login UI (with a `flow_id` session cookie) — or, when
            a live session satisfies `prompt`/`max_age`/`id_token_hint`,
            straight back to `redirect_uri?code=…&state=…` with
            no login surface at all. A post-validation error instead redirects to
            `redirect_uri?error=<code>&state=<state>` (`invalid_scope`,
            `invalid_target`, or `login_required` for a failed `prompt=none`).
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              description:
                Short-lived HttpOnly Secure SameSite=Lax flow_id cookie.
              schema:
                type: string
        "400":
          description: |
            invalid_request / unsupported_response_type / invalid client or
            redirect_uri — rendered on the service error page (not redirected).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthError"
...
components:
  schemas:
...
    OAuthError:
      type: object
      description:
        OAuth 2.0 error response. Distinct from the API's generic `{message}`
        errors.
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - unauthorized_client
            - unsupported_grant_type
            - invalid_scope
            - invalid_target
            - invalid_token
          example: "invalid_grant"
        error_description:
          type: string
          description:
            Human-readable detail. Does not distinguish why the token was
            rejected.
          example: "The refresh token is invalid, expired, or revoked."
...
  securitySchemes:
    CognitoAuthorizer:
      type: apiKey
      name: Authorization
      in: header
      description: |
        Admin Cognito User Pool JWT, for the admin APIs (`/v1/admin/*`).
        End-user routes take an ESP User token instead — see `bearerAuth`.
      x-amazon-apigateway-authtype: cognito_user_pools
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An ESP User access token, sent as a Bearer token. The token is verified
        against the provider's published public keys.
    clientBasicAuth:
      type: http
      scheme: basic
      description: |
        OAuth client authentication (RFC 6749 §2.3.1) as HTTP Basic
        `base64(client_id:client_secret)`. A confidential client sends its secret;
        a public client sends its client_id with an empty secret.
security: []